# GH-2998 native store and LMDB store memory overflow performance