--- a/core/model/src/main/java/org/eclipse/rdf4j/model/impl/LinkedHashModel.java
+++ b/core/model/src/main/java/org/eclipse/rdf4j/model/impl/LinkedHashModel.java
@@ -341,13 +341,13 @@ public class LinkedHashModel extends AbstractModel {
 
 		private static final long serialVersionUID = -1205676084606998540L;
 
-		Set<ModelStatement> subjects = new LinkedHashSet<>();
+		Set<ModelStatement> subjects = new LinkedHashSet<>(1);
 
-		Set<ModelStatement> predicates = new LinkedHashSet<>();
+		Set<ModelStatement> predicates = new LinkedHashSet<>(1);
 
-		Set<ModelStatement> objects = new LinkedHashSet<>();
+		Set<ModelStatement> objects = new LinkedHashSet<>(1);
 
-		Set<ModelStatement> contexts = new LinkedHashSet<>();
+		Set<ModelStatement> contexts = new LinkedHashSet<>(1);
 
 		private final V value;