--- a/core/model/src/main/java/org/eclipse/rdf4j/model/impl/LinkedHashModel.java
+++ b/core/model/src/main/java/org/eclipse/rdf4j/model/impl/LinkedHashModel.java
@@ -420,6 +420,13 @@ public class LinkedHashModel extends AbstractModel {
 			if (this == other) {
 				return true;
 			}
+			if (other == null) {
+				return false;
+			}
+			if (other == statement) {
+				return true;
+			}
+
 			if (!super.equals(other)) {
 				return false;
 			}