--- a/core/sail/lmdb/src/main/java/org/eclipse/rdf4j/sail/lmdb/LmdbSailStore.java
+++ b/core/sail/lmdb/src/main/java/org/eclipse/rdf4j/sail/lmdb/LmdbSailStore.java
@@ -16,10 +16,7 @@ import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.ExecutorService;
@@ -75,7 +72,7 @@ class LmdbSailStore implements SailStore {
 	private volatile boolean asyncTransactionFinished;
 	private volatile boolean nextTransactionAsync;
 
-	private final boolean enableMultiThreading = true;
+	boolean enableMultiThreading = true;
 
 	private PersistentSetFactory<Long> setFactory;
 	private PersistentSet<Long> unusedIds, nextUnusedIds;
@@ -571,6 +568,51 @@ class LmdbSailStore implements SailStore {
 		@Override
 		public void approve(Resource subj, IRI pred, Value obj, Resource ctx) throws SailException {
 			addStatement(subj, pred, obj, explicit, ctx);
+		}
+
+		@Override
+		public void approveAll(Set<Statement> approved, Set<Resource> approvedContexts) {
+
+			sinkStoreAccessLock.lock();
+			try {
+				startTransaction(true);
+
+				for (Statement statement : approved) {
+					Resource subj = statement.getSubject();
+					IRI pred = statement.getPredicate();
+					Value obj = statement.getObject();
+					Resource context = statement.getContext();
+
+					AddQuadOperation q = new AddQuadOperation();
+					q.s = valueStore.storeValue(subj);
+					q.p = valueStore.storeValue(pred);
+					q.o = valueStore.storeValue(obj);
+					q.c = context == null ? 0 : valueStore.storeValue(context);
+					q.context = context;
+					q.explicit = explicit;
+
+					if (multiThreadingActive) {
+						while (!opQueue.add(q)) {
+							if (tripleStoreException != null) {
+								throw wrapTripleStoreException();
+							}
+						}
+
+					} else {
+						q.execute();
+					}
+
+				}
+			} catch (IOException e) {
+				rollback();
+				throw new SailException(e);
+			} catch (RuntimeException e) {
+				rollback();
+				logger.error("Encountered an unexpected problem while trying to add a statement", e);
+				throw e;
+			} finally {
+				sinkStoreAccessLock.unlock();
+			}
 		}
 
 		@Override