--- a/core/sail/lmdb/src/main/java/org/eclipse/rdf4j/sail/lmdb/LmdbSailStore.java
+++ b/core/sail/lmdb/src/main/java/org/eclipse/rdf4j/sail/lmdb/LmdbSailStore.java
@@ -596,7 +596,7 @@ class LmdbSailStore implements SailStore {
 							if (tripleStoreException != null) {
 								throw wrapTripleStoreException();
 							}
-							Thread.yield();
+							Thread.onSpinWait();
 						}
 
 					} else {
@@ -733,7 +733,7 @@ class LmdbSailStore implements SailStore {
 						if (tripleStoreException != null) {
 							throw wrapTripleStoreException();
 						} else {
-							Thread.yield();
+							Thread.onSpinWait();
 						}
 					}
 				} else {