@hmottestad 2024-05-09
```
# JMH version: 1.37
# VM version: JDK 22, OpenJDK 64-Bit Server VM, 22+36
# VM invoker: /Library/Java/JavaVirtualMachines/zulu-22.jdk/Contents/Home/bin/java
# VM options: -Xms500M -Xmx500M -XX:+UseParallelGC
# Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
# Warmup: <none>
# Measurement: 10 iterations, 1 ms each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Average time, time/op
# Benchmark: org.eclipse.rdf4j.sail.nativerdf.benchmark.OverflowBenchmarkReal.loadLotsOfData

# Run progress: 0.00% complete, ETA 00:00:00
# Fork: 1 of 1
Iteration   1: 13:14:30.942 [org.eclipse.rdf4j.sail.nativerdf.benchmark.OverflowBenchmarkReal.loadLotsOfData-jmh-worker-1] DEBUG syncing at 450000 triples. max block size: 62357688
13:14:30.944 [org.eclipse.rdf4j.sail.nativerdf.benchmark.OverflowBenchmarkReal.loadLotsOfData-jmh-worker-1] DEBUG memory overflow using temp directory /var/folders/0m/x3jzy4t15djd88znchzzx1_h0000gn/T/model12752373344023074604
13:15:03.288 [org.eclipse.rdf4j.sail.nativerdf.benchmark.OverflowBenchmarkReal.loadLotsOfData-jmh-worker-1] DEBUG overflow synced to disk
49493.922 ms/op
Iteration   2: 13:15:26.215 [org.eclipse.rdf4j.sail.nativerdf.benchmark.OverflowBenchmarkReal.loadLotsOfData-jmh-worker-1] WARN  Rolling back transaction due to connection close
<failure>

java.lang.OutOfMemoryError: GC overhead limit exceeded
	at java.base/java.util.Arrays.copyOfRange(Arrays.java:3852)
	at java.base/java.lang.StringLatin1.newString(StringLatin1.java:864)
	at java.base/java.lang.String.substring(String.java:2881)
	at org.eclipse.rdf4j.common.net.ParsedIRI.parseMember(ParsedIRI.java:987)
	at org.eclipse.rdf4j.common.net.ParsedIRI.parseScheme(ParsedIRI.java:848)
	at org.eclipse.rdf4j.common.net.ParsedIRI.parse(ParsedIRI.java:771)
	at org.eclipse.rdf4j.common.net.ParsedIRI.<init>(ParsedIRI.java:268)
	at org.eclipse.rdf4j.rio.helpers.AbstractRDFParser.createURI(AbstractRDFParser.java:351)
	at org.eclipse.rdf4j.rio.helpers.AbstractRDFParser.resolveURI(AbstractRDFParser.java:341)
	at org.eclipse.rdf4j.rio.turtle.TurtleParser.parseURI(TurtleParser.java:944)
	at org.eclipse.rdf4j.rio.turtle.TurtleParser.parseValue(TurtleParser.java:571)
	at org.eclipse.rdf4j.rio.turtle.TurtleParser.parseSubject(TurtleParser.java:398)
	at org.eclipse.rdf4j.rio.turtle.TurtleParser.parseTriples(TurtleParser.java:333)
	at org.eclipse.rdf4j.rio.turtle.TurtleParser.parseStatement(TurtleParser.java:203)
	at org.eclipse.rdf4j.rio.turtle.TurtleParser.parse(TurtleParser.java:165)
	at org.eclipse.rdf4j.rio.turtle.TurtleParser.parse(TurtleParser.java:128)
	at org.eclipse.rdf4j.repository.util.RDFLoader.loadInputStreamOrReader(RDFLoader.java:302)
	at org.eclipse.rdf4j.repository.util.RDFLoader.load(RDFLoader.java:229)
	at org.eclipse.rdf4j.repository.util.RDFLoader.load(RDFLoader.java:191)
	at org.eclipse.rdf4j.repository.base.AbstractRepositoryConnection.add(AbstractRepositoryConnection.java:257)
	at org.eclipse.rdf4j.repository.RepositoryConnection.add(RepositoryConnection.java:785)
	at org.eclipse.rdf4j.sail.nativerdf.benchmark.OverflowBenchmarkReal.loadLotsOfData(OverflowBenchmarkReal.java:79)
	at org.eclipse.rdf4j.sail.nativerdf.benchmark.jmh_generated.OverflowBenchmarkReal_loadLotsOfData_jmhTest.loadLotsOfData_avgt_jmhStub(OverflowBenchmarkReal_loadLotsOfData_jmhTest.java:236)
	at org.eclipse.rdf4j.sail.nativerdf.benchmark.jmh_generated.OverflowBenchmarkReal_loadLotsOfData_jmhTest.loadLotsOfData_AverageTime(OverflowBenchmarkReal_loadLotsOfData_jmhTest.java:176)
	at java.base/java.lang.invoke.DirectMethodHandle$Holder.invokeSpecial(DirectMethodHandle$Holder)
	at java.base/java.lang.invoke.LambdaForm$MH/0x0000038001020800.invoke(LambdaForm$MH)
	at java.base/java.lang.invoke.LambdaForm$MH/0x0000038001020c00.invokeExact_MT(LambdaForm$MH)
	at java.base/jdk.internal.reflect.DirectMethodHandleAccessor.invokeImpl(DirectMethodHandleAccessor.java:155)
	at java.base/jdk.internal.reflect.DirectMethodHandleAccessor.invoke(DirectMethodHandleAccessor.java:103)
	at java.base/java.lang.reflect.Method.invoke(Method.java:580)
	at org.openjdk.jmh.runner.BenchmarkHandler$BenchmarkTask.call(BenchmarkHandler.java:527)
	at org.openjdk.jmh.runner.BenchmarkHandler$BenchmarkTask.call(BenchmarkHandler.java:504)




Result "org.eclipse.rdf4j.sail.nativerdf.benchmark.OverflowBenchmarkReal.loadLotsOfData":
  49493.922 ms/op


# Run complete. Total time: 00:00:57

REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
experiments, perform baseline and negative tests that provide experimental control, make sure
the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
Do not assume the numbers tell you what you want them to tell.

NOTE: Current JVM experimentally supports Compiler Blackholes, and they are in use. Please exercise
extra caution when trusting the results, look into the generated code to check the benchmark still
works, and factor in a small probability of new VM bugs. Additionally, while comparisons between
different JVMs are already problematic, the performance difference caused by different Blackhole
modes can be very significant. Please make sure you use the consistent Blackhole mode for comparisons.

Benchmark                             Mode  Cnt      Score   Error  Units
OverflowBenchmarkReal.loadLotsOfData  avgt       49493.922          ms/op

Process finished with exit code 0
```