@hmottestad 2024-05-14
# LMDB

## Before

## After

```
Benchmark                                                      Mode  Cnt      Score      Error  Units
TransactionsPerSecondBenchmarkFoaf.transaction100kx           thrpt    5      3.968 ±    0.187  ops/s
TransactionsPerSecondBenchmarkFoaf.transaction100kxLevelNone  thrpt    5      5.513 ±    0.213  ops/s
TransactionsPerSecondBenchmarkFoaf.transaction10kx            thrpt    5     38.998 ±    2.844  ops/s
TransactionsPerSecondBenchmarkFoaf.transaction10kxLevelNone   thrpt    5     47.498 ±    4.477  ops/s
TransactionsPerSecondBenchmarkFoaf.transaction10x             thrpt    5  12167.252 ± 1844.259  ops/s
TransactionsPerSecondBenchmarkFoaf.transaction10xLevelNone    thrpt    5  12603.226 ± 1229.140  ops/s
TransactionsPerSecondBenchmarkFoaf.transaction1x              thrpt    5  22407.261 ± 1562.692  ops/s
TransactionsPerSecondBenchmarkFoaf.transaction1xLevelNone     thrpt    5  25590.084 ± 1507.929  ops/s
```